# Refactor SMTSolver and improve UFTheory performance