This PR includes two main changes:

* The `SMTSolver` interface is refactored. The separate `rejectBooleanModel` and `learnAdditional` methods have been merged into a single `learnTheoryClauses`
* The `UFTheory` implementation has been updated to avoid creating a new equality expression ` b == a` when a structure `a == b` already exists.
   * This has a noticeable performance increase for certain tests:
      * current main e332c40bcc63d80d11aa70026baaf9faee260257
          > Slowest: verify_tests.duplicated-assert-is-fast took 1776 ms
          > Done in 4.55s. 
      * d2f0b31
          > Slowest: verify_tests.duplicated-assert-is-fast took 1259 ms
          > Done in 3.13s.